+++ src/components/misc/index.ts
@@ -3,10 +3,12 @@ import StudyList from './StudyList';
 import SurveyList from './SurveyList';
 import RequiredSurveys from './RequiredSurveys';
 import OptionalSurveys from './OptionalSurveys';
+import ReportList from './ReportList';
 
 export {
   LinkList,
   StudyList,
+  ReportList,
   RequiredSurveys,
   OptionalSurveys,
   SurveyList,